refactor: 이메일 디렉터리 추가, 구조 변경
