+++ src/main/java/com/lubycon/curriculum/email/domain/EmailTemplate.java
@@ -1,5 +1,6 @@
-package com.lubycon.curriculum.subscribe.domain;
+package com.lubycon.curriculum.email.domain;
 
+import java.time.LocalDateTime;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -12,18 +13,31 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
 @Entity
-public class Tester {
+public class EmailTemplate {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Column(name = "id", updatable = false)
   private Long id;
 
-  @Column(name = "email", nullable = false, unique = true)
-  private String email;
+  @Column(name = "subject")
+  private String subject;
 
-  public Tester(final String email) {
-    this.email = email;
+  @Column(name = "url")
+  private String url;
+
+  @Column(name = "is_send")
+  private boolean send;
+
+  @Column(name = "created_at", nullable = false)
+  private LocalDateTime createdAt;
+
+  public boolean isAlreadySent() {
+    return send;
+  }
+
+  public void sendComplete() {
+    this.send = true;
   }
 
 }
