+++ src/main/java/com/lubycon/curriculum/email/domain/Tester.java
@@ -1,0 +1,29 @@
+package com.lubycon.curriculum.email.domain;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+@Entity
+public class Tester {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  @Column(name = "id", updatable = false)
+  private Long id;
+
+  @Column(name = "email", nullable = false, unique = true)
+  private String email;
+
+  public Tester(final String email) {
+    this.email = email;
+  }
+
+}
