+++ src/main/java/com/lubycon/curriculum/email/dto/SendMailToSubscribersRequest.java
@@ -1,4 +1,4 @@
-package com.lubycon.curriculum.subscribe.dto;
+package com.lubycon.curriculum.email.dto;
 
 import lombok.Getter;
 import lombok.Setter;
