+++ src/main/java/com/lubycon/curriculum/email/repository/EmailTemplateRepository.java
@@ -1,11 +1,11 @@
-package com.lubycon.curriculum.subscribe.repository;
+package com.lubycon.curriculum.email.repository;
 
-import com.lubycon.curriculum.subscribe.domain.EmailTemplate;
+import com.lubycon.curriculum.email.domain.EmailTemplate;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EmailTemplateRepository extends JpaRepository<EmailTemplate, Long> {
 
   Optional<EmailTemplate> findFirstByOrderByCreatedAtDesc();
-  
+
 }
