+++ src/main/java/com/lubycon/curriculum/email/repository/TesterRepository.java
@@ -1,6 +1,6 @@
-package com.lubycon.curriculum.subscribe.repository;
+package com.lubycon.curriculum.email.repository;
 
-import com.lubycon.curriculum.subscribe.domain.Tester;
+import com.lubycon.curriculum.email.domain.Tester;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TesterRepository extends JpaRepository<Tester, Long> {
