+++ src/main/java/com/lubycon/curriculum/email/service/EmailTesterService.java
@@ -1,8 +1,8 @@
-package com.lubycon.curriculum.subscribe.service;
+package com.lubycon.curriculum.email.service;
 
-import com.lubycon.curriculum.subscribe.domain.EmailTemplate;
-import com.lubycon.curriculum.subscribe.domain.Tester;
-import com.lubycon.curriculum.subscribe.repository.TesterRepository;
+import com.lubycon.curriculum.email.domain.EmailTemplate;
+import com.lubycon.curriculum.email.domain.Tester;
+import com.lubycon.curriculum.email.repository.TesterRepository;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
@@ -14,7 +14,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class EmailTesterService {
 
   private final TesterRepository testerRepository;
-  private final SendEmailService sendEmailService;
+  private final CourseEmailService sendEmailService;
   private final EmailTemplateService emailTemplateService;
 
   @Transactional
